affine textures, (very bad) linear programming
git-svn-id: svn://svn.code.sf.net/p/libsh/code/sh/branches/arithmetic@1969 afdca40c-03d6-0310-8ede-e9f093b21075

+++ src/ShRecord.cpp
@@ -71,7 +71,7 @@ ShRecord::VarList::iterator ShRecord::end()
   return m_vars.end();
 }
 
-int ShRecord::size() const 
+size_t ShRecord::size() const 
 {
   return m_vars.size();
 }
